change gain to int8_t and a few tweeks

+++ components/tas58xx/tas58xx_helpers.h
@@ -14,17 +14,17 @@ struct BiquadCoefficients {
 
 int32_t gain_to_f9_23_(int8_t gain);
 
-BiquadCoefficients equalizer_qfactor_(uint32_t sample_rate, uint16_t frequency, int16_t gain, float q_factor);
+BiquadCoefficients equalizer_qfactor_(uint32_t sample_rate, uint16_t frequency, int8_t gain, float q_factor);
 
-BiquadCoefficients low_shelf_filter_(uint32_t sample_rate, uint16_t frequency, int16_t gain, float q_factor);
+BiquadCoefficients low_shelf_filter_(uint32_t sample_rate, uint16_t frequency, int8_t gain, float q_factor);
 
-BiquadCoefficients high_shelf_filter_(uint32_t sample_rate, uint16_t frequency, int16_t gain, float q_factor);
+BiquadCoefficients high_shelf_filter_(uint32_t sample_rate, uint16_t frequency, int8_t gain, float q_factor);
 
-BiquadCoefficients low_pass_filter_(uint32_t sample_rate, uint16_t frequency, int16_t gain);
+BiquadCoefficients low_pass_filter_(uint32_t sample_rate, uint16_t frequency, int8_t gain);
 
-BiquadCoefficients high_pass_filter_(uint32_t sample_rate, uint16_t frequency, int16_t gain);
+BiquadCoefficients high_pass_filter_(uint32_t sample_rate, uint16_t frequency, int8_t gain);
 
-BiquadCoefficients peaking_eq_(uint32_t sample_rate, uint16_t frequency, int16_t gain, float q_factor);
+BiquadCoefficients peaking_eq_(uint32_t sample_rate, uint16_t frequency, int8_t gain, float q_factor);
 
 BiquadCoefficients band_pass_filter_(uint32_t sample_rate, uint16_t frequency, uint16_t bandwidth);
